PR feedback

+++ scripts/docker-entrypoint.sh
@@ -11,7 +11,11 @@ export QGIS_SERVER_PARALLEL_RENDERING=1
 # Start
 cd /code/g3w-admin
 
-rm -rf /shared-volume/build_done
+# When building in dev env you might want a clean build each time.
+if [[ -z "${DEV}" ]]; then
+  rm -rf /shared-volume/build_done
+fi
+
 
 # Activate the front end app settings
